UI updates

+++ trails/api/management/commands/import_data.py
@@ -6,23 +6,37 @@ from typing import Dict
 
 import djclick as click
 
-from api.models import TrailNetwork, Route, Trailhead, Node
-from osm.loader import OSMIngestor, IngestSettings, DefaultQualitySettings
+from api.models import TrailNetwork, Route, Trailhead, Node, TravelCache
+from osm.loader import OSMIngestor, IngestSettings, DefaultQualitySettings, LocationFilter, OsmiumTrailLoader
 from tqdm import tqdm
 
-Settings = IngestSettings(
-    max_distance_km=50,
-    max_segments=50,
-    max_concurrent=100,
-    quality_settings=DefaultQualitySettings,
-)
 
 
 @click.command()
-@click.option("--parallelism", "-p", type=click.INT)
+@click.option("--parallelism", "-p", type=click.INT, default=multiprocessing.cpu_count())
+@click.option('--center', type=click.STRING)
+@click.option('--radius', type=click.INT)
+@click.option('--reset', type=click.BOOL, help='Delete all data before importing', default=False)
 @click.argument("file", type=click.Path(exists=True))
-def import_data(file: str, parallelism=multiprocessing.cpu_count()):
+def import_data(file: str, center, radius, parallelism, reset):
     start_time = time.time()
+    if center:
+        lat, lon = center.split(',')
+        if radius is None:
+            click.secho('Radius must be specified with lat/lon', fg='red')
+            exit(1)
+        location_filter = LocationFilter(float(lat), float(lon), radius_km=radius)
+    else:
+        location_filter = None
+
+    Settings = IngestSettings(
+        max_distance_km=50,
+        max_segments=300,
+        max_concurrent=40,
+        quality_settings=DefaultQualitySettings,
+        location_filter=location_filter
+    )
+
     path = Path(file)
     loader = OSMIngestor(Settings)
     loader.ingest_file(path, parallelism=parallelism)
@@ -32,11 +46,13 @@ def import_data(file: str, parallelism=multiprocessing.cpu_count()):
         f"OSM data [loops: {result.total_loops()}, trail networks: {len(result.loops.keys())}] ingested in {ingest_time-start_time} seconds"
     )
 
-    # Need to delete everything before import TODO
-    Trailhead.objects.all().delete()
-    Route.objects.all().delete()
-    TrailNetwork.objects.all().delete()
-    Node.objects.all().delete()
+    if reset:
+        # Need to delete everything before import TODO
+        Trailhead.objects.all().delete()
+        Route.objects.all().delete()
+        TrailNetwork.objects.all().delete()
+        Node.objects.all().delete()
+        TravelCache.objects.all().delete()
 
     trailheads: Dict[int, Trailhead] = {}
     for trail_network_osm, loops in tqdm(result.loops.items()):
